Make custom labels and folders selecteable in the sidebar
Now all locations in the sidebar that should show a "selected" state
when clicked do so.
Some of the locations, such as "settings", "subscriptions" or "logout"
do not handle selection as it doesn't make sense: when they are clicked
they do redirect to the needed screen but the sidebar is not visible anymore.

MAILANDR-67