Solve UI test failure due to an issue with string resources
It seems that Context instances accessed in UI tests through `
InstrumentationRegistry.getInstrumentation().getTargetContext()` or `
InstrumentationRegistry.getInstrumentation().getTargetContext().getApplicationContext()`
are not properly updated when the language is changed during the app
runtime.
This causes failures as the app is actually in a different language than the
one that calls to `getString(R.string.key)` return, causing UI elements not be
found.