Define current time zone for FormatMailboxItemTimeTest
Since such tests depend on the creation of Calendar instances, which
implicitly take into consideration the current time zone (see docs for
`Calendar.getInstance`) we need to provide a specific timezone for the
tests, so that the hardcoded times we expect in some of them (eg. 18:40)
can be verified.

MAILANDR-70