Mailbox item mapper explicitly checks for type when mapping icons
As a presentation rule, we never display reply, reply all or forward
icons for "conversation" type items.
This was previously enforced through the MailboxItem itself being
created with all of the above flags hardcoded to false, which then
allowed us to just map it to the Ui model here (mailboxitem.isReplied...).

As the wanted behavior (as described in the test) read "mailbox items of
conversation type never show any of reply, reply-all, forwarded icon",
we now explicitly check the type to be more truthful.

MAILANDR-70