Ignore all-mail and all-sent labels when checking for trash/spam conversation
Most conversations contain all-mail / all-sent labels in their
`labelIds` list. When checking all the labels to define whether all
messages are in either trash or spam, we do not care about the former
ones as they are also applied to trashed messages.

A more robust solution will be applied here after MAILANDR-206 is
completed as commented in the @Ignored test.

MAILANDR-221