Message detail screen reacts to mark read action outcome
The mark unread use case is currently mocked and hardwired to fail.
When the use case returns success, the message detail screen will be
dismissed. When error is returned, it will be shown in a snackbar.

The two "effects" to dismiss / show error were added in the main
`MessageDetailState`. They are currently being assigned with "ad-hoc"
checks done in the VM, a "main" reducer should be extracted to take care
of this instead (as already done in convesationVM)

MAILANDR-220