Make star / unstar uses cases suspend instead of returning flow
The idea behind a flow was to be able to observe both the outcome of the
local operation and the remote one.
This wasn't a good approach, as the local operation should be optimistic
and the remote one should be handled through a separated logical flow so
that it works also when the user is not on a given screen anymore (or
not in the app alltogether).

MAILANDR-225