Adds Immutable annotation to the MailboxItemUiModel
Marking this class as immutable explicitly since the compiler does not consider
it as such. Even though it's a data class, it has lists as params, which
are not considered immutable by compose. It also takes UserId and ConversationId
types as params: both are coming from non-compose modules, and as such
are not considered stable by the compiler.

MAILANDR-268