Do not hide conversation metadata when offline
When opening a conversation while offline, its metadata would be shown
from local DB (always available as it's the same data needed to display
the mailbox).
When the remote request to fetch the converastion details (list of
messages) fails due to being offline, the state of the conversation
would be set to "error", causing the already displayed data to be
hidden.

This is addressed by not changing the conversation's metadata state when
reciving the offline event, based on the expectation that data needs to
be already shown. This can be extended to add dedicated behaviors for
edge cases (ie. data was loading and we get offline event..)

MAILANDR-306