Mark conversation unread marks the latest read message in the current label
Current label being the label this conversation was opened from.

In the MarkConversationUnread use case, we get the label that's
currently selected in the drawer (which we always expect to be the one
we go back to when conversation is closed as a consequence of user
marking it unread) and pass it thorough so that it can be used to apply
the right logic on all levels. (eg. Choosing which message to mark unread,
which conversation's "context label" to increment unreadCount for and
finally passing it as param to the API request so that same logic can be
applied there)

MAILANDR-220