MR comments - Added the offline error message when trying to open a message.
The feedback we give to the user in that case seems to be too slow though and it takes a few seconds for the error message to appear. We could either have a partially loading indicator for the message and then collapse it again if the user is offline, or have a general app wide indicator that the user is offline, and keep the current messages.

MAILANDR-230