Crash when tapping a message when WebView is not installed on device
Added a new device capabilities interface that we can use to inform of different capabilities to other parts of the app. In this change only the ability to use the google webview is implemented.

As we are interested in using this from compose, in the activity we inject the `DeviceCapabilities` interface and pass it to the composition tree using the `CompositionLocalProvider`. This way we can recover the value later in any composable. The alternative to this is passing the value through a view model to reach the composable but that requires more file changes in all the layers just to pass one value.

Added a dagger module in the mail common data package so that we can keep the `DeviceCapabilitiesImpl` internal.

The UI tests only mock the capabilities result. The WebView must be available in the target device when running them.

MAILANDR-451