Mailbox screen displays error information for offline failures
This allows the user to get better feedback on what went wrong,
while doing a mailbox "refresh" operation (which could be a first load
  or a "pull to refresh").

This is achieved by passing the `DataError` encoutered by the mediator
as part of the exception. Any errors other than "offline" are currently
treated the same way (generic request failure) while any error that is
not bringing along explicit failure information (Thus, everything that's
not a DataErrorException) will show an "unexpected error" message.

MAILANDR-561