Only display offline/error with data effects once
When a request to refresh data in mailbox fails due to any error (or
missing network) while some data is shown, we want to show
a snackbar to notify the user while letting the current data on screen.

This error should be shown again only if the user performs a pull to
refresh. Given the state that let us know of such error is part of the
"lazyListState" built by the paging library, it isn't persisted across
recompositions and it would cause any effects to be shown over and over
again.
To solve this issue, we pass such states as if they were user actions
(eg. OnOfflineWithData, OnErrorWithData) to the VM which will then
create effects in the state by also taking into consideration whether a
refresh happened (which allows us to skip recompositions)

MAILANDR-561