Chips email validation, fix unfocused single line
When adding long items to the recipients fields, the counter chip was being put in a second line and the email chip was occupying the complete field.

In this change, we remodel the items that we provide from the state holder and change the composable to accommodate all the possible focused and unfocused states with the single line restriction for the unfocused one.

MAILANDR-587