Chips email validation, fix duplicate emails per field
When adding recipients to any of the composer fields, we need to prevent adding duplicates in the same field, removing the last entered one if duplicate and showing an error message.

As we are now overwriting the recipients in case of this error, the ChipsListState has been modified to instead of accepting an initial value, being able to receive the complete list of emails on demand.

MAILANDR-625