Request notification permissions
After an account is logged in, request the notification permission if it is not given and the device is running Android 13+.

If the user has been asked a few times and denied the permission, the permissions pop up may not appear as it is controlled by the system. In that case we would need a custom app setting to take the user to the app permissions screen so that it can be enabled from there but this is for a separate `Settings` task.

Added exclusion to StrictModeInitializer to prevent a warning that shows when requesting the permission.

MAILANDR-778