Do not sync draft on API when body is empty
The API doesn't accept being called with an empty body.
Body is empty when the draft was just created and any data (subject, recipients..) was added
but the body wasn't. In order to avoid adding ad-hoc logic to create an encrypted empty-string
body, we block the draft creation (triggered by the automatic sync) from happening till a body was added

MAILANDR-491