Ensure latest content inserted by the user is uploaded when composer is closed
We achieve this by "forcing" the upload when user exists the composer.
Forcing is done through appending an upload job to any existing upload
one (compared to the automatic upload logic which would do nothing when
there is an exisiting job ongoing).

This avoids the edge case of not uploading the last content the user
typed in case the previoud request was still ongoing when the composer
is closed.

- ComposerVM runs storing locally and syncing the draft in a non-cancellable context
- StoreAllFieldsLocally returns only after all use cases were called. This is very important to avoid race conditions between storing locally and uploading when composer is closed
- Forcing upload of draft uses workManager's APPEND_OR_REPLACE existing work policy as it's a valid case for previous work to have failed (eg. new draft without body, works fails due to "DraftRequestNotPerformed" to satisfy API constraints)

MAILANDR-491