When uploading a draft that was created from web client, perform an update
The logic to discriminate between a local draft and a remote one, relies
on the existence of an "apiMessageId" field in the DraftState. Such
field is stored after the initial draft creation happens, and it ensures
that creation happens only onces and all subsequent operations are
update.
This doesn't work the first time a draft created from web is opened for
editing in the app: we will create a draft state with no "apiMessageId",
and this draft will be created even though it already exists.

One way to ensure it is differentiated, is to also check for the
DraftState's "messageId" field to have a UUID format: if it does, the
draft is probably local (as we do generate messageIds as UUIDs): if not,
it means that at creation time of the first "draftState" the ID was
already an API-defined one which makes the draft "remote".

MAILANDR-495