Wrap all operations related to storing drafts in transactor
This solves a race condition which would result in a draft being
duplicated locally.
The condition was verifying when the logic handling the result
of a remote draft creation came in and performed the update of
the message (changing the "local" id with the new "remote" id defined by
the API as result of the 'create' request) while any other local
storeDraft- operation (eg. StoreDraftWithBody) was ongoing.

Given the structure of storeDraft- operations is to 1) read the draft
from DB, 2) performs needed updates and 3) store it again to DB, if the
aforementioned result came in anytime in between steps 1) and 3), the
execution of step 3) would cause a new message with the "local" id to be
created locally.

This solution avoids the problem by encapsulating operations 1,2,3 in a
single "block" and ensuring it is executed as part of a single
"transaction", disallowing the DB to be edited while it's ongoing. This
ensures data consistenty as no matter the order, it results in data
being updated from both sources.

MAILANDR-491