test: Fix `MessageDetailViewModelTest`.
Change the structure for some tests with Turbine - consume the state
in separate coroutine.

Change the mock in `givenNoLoggedInUser` to return a `MutableStateFlow`,
so that the tested view model doesn't crash in `getMessageBody > primaryUserId.first()`.
Previously, this would also crash, but the exception was only printed out to the console.