Add separated test tags for each snackbar host in Home
The snacbar hosts were split to have one for each type as a workaround
to handle snacks enqueuing, in the scope of showing message sending
feedback. (This workaround is tracked to be improved).

This commit fixes the UI tests which won't find the sbackbar in the
"default" test tag anymore

MAILANDR-614