When composer is opened with an action, copy parent attachments to the new draft
This takes the needed actions to ensure the draft has the needed data
from the parent. When the actions is reply or reply all, this consists
in copying the "inline" attachments of the parent while whent the action
is forward all of the attachments are copied.

Note that with this initial approach, we do not re-encrypt the
keypackets of any attachments. This still works because usually the
addressKey the attachments were encrypted to in the parent message is
the same one we use for sending: it won't be the case in some edge cases
(eg. message received to pm.me, free user, can't reply using pm.me), for
which a use case to re-encrypt key packets will follow.

MAILANDR-690