Do not drop blank lines when sending a plain text message
They are dropped if we default to "HTML" mimeType (which is the default
for Proton user's send preferences) when a message was written in plain
text. As there's no additional value in wrapping this plaintext with
any html (eg. <pre> tag) we change the mime to plaintext independently
from the preference instead.

MAILANDR-1024