Copy parent attachments to draft when replying to/forwarding a PGP/MIME message
The parent PGP/MIME message has its attachments stored to the cached local storage.
When replying or forwarding we need to copy these attachments to the draft and
set their sync state to Local, so that they can be uploaded before the user
tries to send or when exiting the composer.

MAILANDR-1008