Message Sending failure is notified to the user when happing in early stages
Sending a message usually consist of 3 stages: uplading a draft,
uploading attachments and sending. If any of the first two stages fail
while a message is being sent, we now notify the user of such failure
and move the message back to drafts.

The notification is done through setting the "Draft state" of this
message to "Error Sending" (instead of "Error Uploading Draft /
Attachments" as would be the case if such stages were failing while a
message was not being sent) which is observed higher up in the Compose
hierarchy and will be notified.

MAILANDR-951