Await 5 seconds before starting send message requests sequence
This is a mitigation for an issue that's (inconsistently) seen when a message
is sent while offline. When the device comes back online in a situation
where both LTE and Wi-Fi connectivity are present, it will immediately
connect to LTE network, which will satisfy the "connected" constraint of
the send workers and let them start.
When the device switches to Wi-Fi after few seconds, the worker that's
running (while testing without attachments, SendMessage most of the time)
will be stopped (with reason `STOP_REASON_CONSTRAINT_CONNECTIVITY`) and
then run again later. Whenever this happens after the send message POST
HTTP request was fired, with most probablilty the next instance of the
worker will fail due to the message being already sent (for which API
returns an error), causing a failure message to be shown to the user
while the message was actually sent.

Further details in the task.

MAILANDR-1143