Reuse same send-message packages for different recipients if possible
Instead of GenerateSendMessagePackage creating a package for each recipient individually, we reuse them if possible and put the data specific to recipients into "addresses" map. Packages of type Internal and ExternalCleartext can be shared amongst different recipients of both types. Packages of type ExternalSigned can be shared but only amongst recipients of that type. Packages of type ExternalEncrypted cannot be shared so for them the behavior stays like before.

MAILANDR-946