Fix failure replying to message with embedded images
When having conversation mode enabled and replying to a message with
embedded images, the reply was (inconsistently) failing.
The flow / root cause of the issue was as follows:
- Open conversation with embedded images
- Await for the message to load and hit reply
- [composer opens] type something in the body and hit send

Hitting send triggers the followin ops:
- saveDraftWithAllFields is called, which saves in DB the current content
of composer: recipient, subject, body and embedded images
- Some of this content (namely, the recipient) is not yet uploaded to API,
as the automatic upload (which happens while composer is opened) only
uploads what's in DB _and the prefilled recipient is not stored to DB
till send / close action is triggered_ (due to some issues caused by
triggering save without user interaction)
- the send message flow is enqueued (including upload draft, attachments
  and send calls)
- The composer is dismissed and user goes back to the conversation screen
- Conversation screen re-loads the messages **and in the process of
  marking them as read performs a call to the /conversation/{id}
  endpoint**
- Such call returns the conversation metadata and all its messages,
  **including a not-yet-updated version of the current message** (not
  having the recipients)
- The messages are saved to DB causing the message scheduled to be sent
  to loose the recipient
- Sending fails

This fix introduces a side effect of not seeing immediately the draft /
sent message if staying in the conversation screen right after replying:
this should be addressed by making the local data right (ie. adding
conversation ID to the message that was just sent timely, )
instead of by performing a full-refresh. This should be addressed by
MAILANDR-982.

MAILANDR-1316