Update message unread counters optimistically for inclusive labels as well
When a message that has a custom label or is starred is updated (ie.
marked read / unread, moved or deleted) we update the unread counters
for all of its labels so that inclusive locations (eg. star) are also updated
This will take care of updating all-mail location as well.

MAILANDR-1093