Avoid pin lock bypass by using split screen on MainActivity task
By using a new task for the pin lock screen, it is possible to bypass it
by selecting the main app task (which in recents shows the mailbox) and
asking android to display it using the split screen functionlity.
In this case, instead or redirecting to pin lock task the split screen
can be interacted with and the lock if effectively bypassed.

MAILANDR-1454