Validate sender address when prefiling composer for a draft action
In case the recipient address of the "parent message" (which is the
message we're performing the action on, ie. the message being replied
to) is not valid as a sender address (eg. disabled or pm.me address
for a free user) then a valid fallback address is resolved and prefilled
in the composer instead.

MAILANDR-1012