Re-encrypt parent attachments when sender address is automatically changed
When prefilling a draft for a reply / fw action, we attempt by default
to re-use the same address that as of the parent message to reply (ie.
if the message was sent to alias-address@proton.me, the reply sender
will be the same).
Sometimes such address can't be used for replying, in which cases it's
switched to the user's default address instead. When this happens,
parent message attachments needs to be re-encrypted with the new
address.

MAILANDR-1012