Do not crash when opening a message sent while offline
This is strictly related to the context explained in the previous commit.
When having conversation mode enabled and sending a message offline, the
message will be enqueued and moved to "sent" folder.
Till it gets actually created on the backend, it will not have a
`conversationId`, which causes an IllegalState when the user opens it
(as we try to navigate to an empty conversationId).

This is solved by falling back to opening it as a message when in such
case.

MAILANDR-982