Correct date formatting issue for some messages from previous years
For messages with same week number as current but from a previous year
the isCurrentWeek logic was followed, as it was not taking into
consideration the year but only the week-of-year.
The fix adds consideration for the year is "isCurrentWeek" and one
additional method to still cover the case of the same week across year
change.

MAILANDR-925