Respect the correct MIME type when sending a message
When opening a draft in Composer that has been created on web, if the mime type is HTML we need to keep it instead of reverting to Plain Text. If we fail to do so, when the message is sent all the new lines will be lost.

The would happen when replying to an HTML message from the app: the original mime type of the draft is HTML, and in this case we need to convert our draft body to HTML instead of wrapping it into a <pre> tag.

MAILANDR-1108