Adapt css injection to the WebView theme preference
Add the media query only when the app needs to follow the theme of the app/system
and when the WebView theme preference is set to dark. This is needed because the
media query caused problems in the scenario where the app/system theme is dark and
the WebView theme preference is set to light.

MAILANDR-1570