Avoid race condition with API calls when undoing a move operation
When performing a move operation (while offline in particular, but not only
and undoing it through the "undo" button (or manually, by going to the
destination folder and performing the opposite action), the API calls
that are fired by the workers (in case of the op being performed offline
once back online) can run and complete in any order: when the order is
not the scheduling one, the undo of the operation silently fails.

This can be prevented by forcing the api calls that are performing the
same type of operation (eg. add label) to happen sequentially as
scheduled.

MAILANDR-1727