Keep cursor into view while in Composer
Make sure the cursor is always visible when the user types anything in the message body. This is a workaround and requires moving from Material3 TextField to BasicTextField and tweaking a bit the behaviour with BringIntoViewRequester.

Starting from 1.7.0 we should aim to explore BTF2 and remove the workaround while providing the same user experience.

MAILANDR-2096