Fix crash when replying to or forwarding a very big message
We store big messages in files instead of the DB. When the draft is
created locally for the big message, it will be saved to a file which
will have the local message id as name. After the draft has been
created on the BE, we need to update the file name to the remote
message id.

MAILANDR-2062