Clear focus when showing the upselling bottom sheet
This is a fix for a behavior where the keyboard would be opened automatically
after initiating the payment process from the upselling bottom sheet.
This happens when a text field already had the focus before opening the bottom sheet.
This fix makes sure that no text field has the focus when opening the bottom sheet,
so no text field can request it before the bottom sheet is closed.

MAILANDR-2212