Update exit actions behaviour in Conversation detail VM
The VM is observing a lot of things, and when several actions are performed in chain (like move + label, or unlabel + relabel) and the user is leaving the screen, sometimes race conditions cause the local data to be left behind compared to the correct state. The state is the fixed upon the next event loop run, but the inconsistency makes the UX suboptimal as it leaves the user wonder why the actions hasn't worked as expected.

When performing actions that are expected to make the user leave the screen, all observers are now stopped and won't be resumed unless the action actually fails. This prevents unnecessary state emissions.

Some parts of the VM have also been cleaned up, as for instance attachments observations that are now consistent and not duplicated (previously collapsing and expanding messages would spawn observers relentlessly).

Part of these changes will need to be also applied to the Message details VM.

MAILANDR-2289