Change the logic for handling the visibility of the bottom sheets
This change mitigates the "jumping" problem of the upselling bottom sheets.
These bottom sheets have separate view models and giving them as little as
100 milliseconds to initialize before calling show() prevents them from "jumping".
Furthermore, changing the value of "showBottomSheet" to true just once when
the bottom sheet is called for the first time is enough to prevent preloading
when the bottom sheet shouldn't be shown.

MAILANDR-2190