Rework message move from three dots icon
The new logic allows better granularity when performing actions on a message level. Labels are applied and removed (when necessary) also to the conversation but only locally as we expect the event loop to align the state next time it runs.

This change is required to make sure that when a message that is located in Folder A is moved out of it and the conversation does not contain any other messages located in Folder A, the starting folder label (unless it's AllMail/AlmostAllMail) is removed and the destination folder is applied also to the conversation, preventing a conversation to stay in a location where it has no messages.

MAILANDR-2288