Enforce queuing on UploadDraft during Sending flow
Introduce another unique `workId` from the Sending flow workers chain and ensure that the first worker is always run AFTER any existing "UploadDraftWorker" job previously enqueued.

The entry point information is passed as tag (we use the uniqueWorkId), which can either be:
- "SyncDraftWorker-${messageId.id}" for the continuous upload work;
- "SendMessageWorker-${messageId.id}" for the upload that is part of the sending chain.

Upon performing `doWork()`, if the entry point is from the sending chain, `WorkManager` is queried to check the status of the original continuous upload work, and depending on its state, that same job is cancelled or the sending job is run immediately.

More in details:
- Worker triggered by continuous upload is ENQUEUED/BLOCKER -> cancel continuous upload work and Result.retry() on sending worker.
- Worker triggered by continuous upload is RUNNING -> do not cancel continuous upload work and Result.retry() on sending worker.
- Worker triggered by continuous upload is SUCCEEDED/FAILED/CANCELLED -> do not cancel continuous upload work, run sending worker work immediately.
- No info given for worker triggered by continuous upload -> do not cancel, run sending worker immediately.

This change is needed to guarantee an APPEND_OR_REPLACE policy while still allowing the sending chain to run if the work from the continuous upload fails. Having then the sending flow an UploadDraftWorker as first dependency of the chain, it is guaranteed that the upload will run with the latest saved data and ensure consistency upon sending.

MAILANDR-2397