Skip the first mediator REFRESH event within RemoteMediator
When navigating through label ids, a change of label (via SelectedMailLabelId) recreates the Pager + Mediator, causing the Mediator to perform an immediate REFRESH which performs an API call and fetches the messages/conversations list.

This should not be necessary as new messages are returned via the event loop, and the call to refresh the whole list should only be performed when the user performs a pull to refresh action.

The feature flag is temporary and scheduled to be removed within the next few releases.

MAILANDR-2365